Move MemberEffects inline in members/handlers.ts
