+++ administration/server/src/api/members/endpoints.ts
@@ -33,18 +33,18 @@ export const membersGroup = HttpApiGroup.make('members')
       )
   )
   .add(
-    HttpApiEndpoint.post('api.members.create', '/api/members')
-      .setPayload(CreateMemberSchema)
-      .addSuccess(MemberSchema, { status: 201 })
-      .addError(UniqueError)
-      .annotate(OpenApi.Description, 'Create a new member')
-  )
-  .add(
     HttpApiEndpoint.get('api.members.read', '/api/members/:id')
       .setPath(Schema.Struct({ id: Schema.NumberFromString }))
       .addSuccess(MemberSchema)
       .addError(NotFoundError)
       .annotate(OpenApi.Description, 'Get a member by ID')
+  )
+  .add(
+    HttpApiEndpoint.post('api.members.create', '/api/members')
+      .setPayload(CreateMemberSchema)
+      .addSuccess(Schema.Void, { status: 201 })
+      .addError(UniqueError)
+      .annotate(OpenApi.Description, 'Create a new member')
   )
   .add(
     HttpApiEndpoint.put('api.members.update', '/api/members/:id')
@@ -59,7 +59,7 @@ export const membersGroup = HttpApiGroup.make('members')
   .add(
     HttpApiEndpoint.del('api.members.delete', '/api/members/:id')
       .setPath(Schema.Struct({ id: Schema.NumberFromString }))
-      .addSuccess(Schema.Struct({ message: Schema.String }))
+      .addSuccess(Schema.Void)
       .addError(NotFoundError)
       .annotate(OpenApi.Description, 'Delete a member')
   )
@@ -67,7 +67,7 @@ export const membersGroup = HttpApiGroup.make('members')
     HttpApiEndpoint.post('api.members.note', '/api/members/:id/note')
       .setPath(Schema.Struct({ id: Schema.NumberFromString }))
       .setPayload(Schema.Struct({ content: Schema.String }))
-      .addSuccess(Schema.Struct({ message: Schema.String }), { status: 201 })
+      .addSuccess(Schema.Void, { status: 201 })
       .addError(NotFoundError)
       .addError(ParseError)
       .annotate(OpenApi.Description, 'Add a note to a member profile')
